requirements.txt
fastapi
pydantic
faiss-cpu
numpy
sentence-transformers
requests
PyMuPDF
tqdm
spacy
streamlit
uvicorn
urllib3
pytesseract
pillow
nltk
pikepdf
pdfplumber
camelot